+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - git: synchronise settings from ~/.dotfiles/config/git.toml
 
+### Removed
+
+- rustc: no longer install "nightly" toolchain by default
+
 ## [0.28.0] - 2019-03-05
 
 ### Added